# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BizTrust brokerage API",
    "version": "0.1.0",
    "description": "Local demonstration contracts. All products, payments and insurer outcomes are synthetic. No actual cover. No live provider integration is configured."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
...
    "/api/catalog": {
      "get": {
        "summary": "Versioned synthetic multi-insurer catalogue",
        "security": [],
        "responses": {
          "200": {
            "description": "Catalogue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "categories",
                    "insurers",
                    "products",
                    "synthetic"
                  ],
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "insurers": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    },
                    "synthetic": {
                      "const": true
                    },
                    "mode": {
                      "const": "demonstration"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/session": {
      "get": {
        "summary": "Current server session, CSRF token and identity capabilities",
        "security": [],
        "responses": {
          "200": {
            "description": "Session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/demo": {
      "post": {
        "summary": "Create isolated customer demo session",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "From GET /api/session; mutations also require an exact same-origin Origin header."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Loopback demonstration only. Rotates the session; user cannot choose tenant or role."
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Revoke local session and create anonymous session",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "From GET /api/session; mutations also require an exact same-origin Origin header."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "get": {
        "summary": "Start Logto-compatible confidential OIDC code flow with state, nonce and PKCE",
        "responses": {
          "302": {
            "description": "Identity provider redirect"
          },
          "503": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/callback": {
      "get": {
        "summary": "Verify OIDC response and establish server-selected membership",
        "responses": {
          "302": {
            "description": "Signed-in application redirect"
          },
          "400": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes": {
      "post": {
        "summary": "Create a reproducible, owned 30-minute quote",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "From GET /api/session; mutations also require an exact same-origin Origin header."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "400": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/applications/{reference}": {
      "parameters": [
        {
          "name": "reference",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^BT-[0-9]{4}-[A-F0-9]{10}$"
          }
        }
      ],
      "get": {
        "summary": "Application, invoice, quote and audit history, subject to ownership",
        "responses": {
          "200": {
            "description": "Application detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "application",
                    "invoice",
                    "quote",
                    "history"
                  ],
                  "properties": {
                    "application": {
                      "type": "object"
                    },
                    "invoice": {
                      "type": "object"
                    },
                    "quote": {
                      "type": "object"
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Typed error with safe message and request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bt_session",
        "description": "Opaque HttpOnly SameSite=Lax server session, Secure on HTTPS."
      },
      "staffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bt_ops_session",
        "description": "Separate HttpOnly staff session, issued only after staff OIDC and pre-provisioned membership."
      },
      "providerSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Provider-Signature",
        "description": "64-character lowercase hex HMAC-SHA256. See callback authentication contract."
      }
    },
    "schemas": {
      "Empty": {
        "type": "object",
        "maxProperties": 0
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "Product": {
        "type": "object",
        "required": [
          "id",
          "category",
          "insurerId",
          "version",
          "ruleVersion",
          "synthetic",
          "coverage",
          "limit",
          "exclusion",
          "deductible",
          "source"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "insurerId": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "ruleVersion": {
            "type": "string"
          },
          "synthetic": {
            "const": true
          },
          "basePremium": {
            "type": "number"
          },
          "period": {
            "enum": [
              "day",
              "year"
            ]
          },
          "coverage": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "limit": {
            "type": "string"
          },
          "exclusion": {
            "type": "string"
          },
          "deductible": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "flex": {
            "type": "object",
            "description": "Synthetic coverage range, increment, base amount and insurer-specific optional-benefit percentage rates. Runtime validates selection against this product configuration."
          }
        }
      },
...
    }
  }
}